+++ src/componentsCodegen/index.ts
@@ -1,24 +1,26 @@
-import { IDefinitions } from '../swaggerInterfaces'
+import { IDefinitions, IComponents } from '../swaggerInterfaces'
 import { refClassName, isGenerics } from '../utils'
 import { createDefinitionClass } from './createDefinitionClass'
 import { createDefinitionEnum } from './createDefinitionEnum'
 import { IDefinitionClasses, IDefinitionEnums } from '../baseInterfaces'
 
-export function definitionsCodeGen(definitions: IDefinitions) {
+export function componentsCodegen(definitions: IComponents) {
   let definitionModels: IDefinitionClasses = {}
   let definitionEnums: IDefinitionEnums = {}
-  for (const [k, v] of Object.entries(definitions)) {
+  if (!definitions || !definitions.schemas) {
+    definitions = {
+      schemas: {}
+    }
+  }
+  for (const [k, v] of Object.entries(definitions.schemas)) {
     let className = refClassName(k)
     // 如果已经转为泛型类型，则不需要重新定义
-    if (isGenerics(className)) {
-      continue
-    }
-
+    if (isGenerics(className)) continue
     let result = null
     // is an enum definition,just in swagger openAPI v2
     if (v.enum) {
       const enumDef = createDefinitionEnum(className, v.enum, v.type)
-      definitionEnums[`#/definitions/${k}`] = {
+      definitionEnums[`#/components/schemas/${k}`] = {
         name: enumDef.name,
         value: enumDef
       }
@@ -32,13 +34,13 @@ export function definitionsCodeGen(definitions: IDefinitions) {
         // definitionModels[item.name] = {
         //   value: item.text
         // }
-        definitionEnums[`#/definitions/${item.name}`] = {
+        definitionEnums[`#/components/schemas/${item.name}`] = {
           name: item.name,
           content: item.text
         }
       })
 
-      definitionModels[`#/definitions/${k}`] = {
+      definitionModels[`#/components/schemas/${k}`] = {
         value: model,
         name: className
       }