+++ src/componentsCodegen/propTrueType.ts
@@ -12,16 +12,16 @@ export function propTrueType(v: IDefinitionProperty): {
     isType: false,
     ref: ''
   }
-  if (v.$ref) {
+  if (v.$ref || (v.allOf && v.allOf[0])) {
     // 是引用类型
-    result.propType = refClassName(v.$ref)
+    result.propType = refClassName(v.$ref || v.allOf[0].$ref)
     result.ref = result.propType
   }
   //是个数组
   else if (v.items) {
-    if (v.items.$ref) {
+    if (v.items.$ref || (v.items.allOf && v.items.allOf[0])) {
       // 是个引用类型
-      result.ref = refClassName(v.items.$ref)
+      result.ref = refClassName(v.items.$ref || v.items.allOf[0].$ref)
       result.propType = result.ref + '[]'
     } else {
       if (v.items.type === "array") {
@@ -39,12 +39,7 @@ export function propTrueType(v: IDefinitionProperty): {
   // 是枚举 并且是字符串类型
   else if (v.enum && v.type === 'string') {
     result.isEnum = true
-    result.propType = getEnums(v.enum).map(item => {
-      if (isNaN(item)){
-        return `'${item}'='${item}'`;
-      }
-      return  `'KEY_${item}'='${item}'`;
-      }).join(',')
+    result.propType = getEnums(v.enum).map(item => `'${item}'='${item}'`).join(',')
   }
   else if (v.enum) {
     result.isType = true