+++ src/index.ts
@@ -1,14 +1,23 @@
 import * as fs from 'fs'
 import * as path from 'path'
-import prettier from 'prettier';
-import axios from 'axios';
+import prettier from 'prettier'
+import axios from 'axios'
 import pascalcase from 'pascalcase'
+import { ISwaggerOptions, IInclude } from './baseInterfaces'
 import { ISwaggerSource } from './swaggerInterfaces'
+import {
+  requestTemplate,
+  serviceTemplate,
+  enumTemplate,
+  interfaceTemplate,
+  classTemplate,
+  typeTemplate
+} from './templates/template'
+import { customerServiceHeader, serviceHeader } from './templates/serviceHeader'
+import { isOpenApi3, findDeepRefs, setDefinedGenericTypes } from './utils'
+import { requestCodegen } from './requestCodegen'
+import { componentsCodegen } from './componentsCodegen'
 import { definitionsCodeGen } from './definitionCodegen'
-import { enumTemplate, typeTemplate, classTemplate, serviceHeader, customerServiceHeader, serviceTemplate, requestTemplate, interfaceTemplate } from './template';
-import { requestCodegen } from './requestCodegen';
-import { ISwaggerOptions, IInclude } from './baseInterfaces';
-import { findDeepRefs } from './utils';
 
 const defaultOptions: ISwaggerOptions = {
   serviceNameSuffix: 'Service',
@@ -22,23 +31,23 @@ const defaultOptions: ISwaggerOptions = {
   include: [],
   strictNullChecks: true,
   useClassTransformer: false,
+  extendGenericType: []
 }
-
 
 export async function codegen(params: ISwaggerOptions) {
   console.time('finish')
   let err
-  let swaggerSource
-
+  let swaggerSource: ISwaggerSource
+  setDefinedGenericTypes(params.extendGenericType)
+  // 获取接口定义文件
   if (params.remoteUrl) {
     const { data: swaggerJson } = await axios({ url: params.remoteUrl, responseType: 'text' })
     if (Object.prototype.toString.call(swaggerJson) === '[object String]') {
-      fs.writeFileSync('./cache_swagger.json', swaggerJson);
-      swaggerSource = require(path.resolve('./cache_swagger.json'));
+      fs.writeFileSync('./cache_swagger.json', swaggerJson)
+      swaggerSource = require(path.resolve('./cache_swagger.json'))
     } else {
       swaggerSource = <ISwaggerSource>swaggerJson
     }
-
   } else if (params.source) {
     swaggerSource = <ISwaggerSource>params.source
   } else {
@@ -49,20 +58,23 @@ export async function codegen(params: ISwaggerOptions) {
     ...defaultOptions,
     ...params
   }
-  let apiSource = options.useCustomerRequestInstance
-    ? customerServiceHeader(options)
-    : serviceHeader(options)
+  let apiSource = options.useCustomerRequestInstance ? customerServiceHeader(options) : serviceHeader(options)
+  // 判断是否是openApi3.0或者swagger3.0
+  const isV3 = isOpenApi3(params.openApi || swaggerSource.openapi || swaggerSource.swagger)
+  console.log('isV3', isV3)
+  let requestClasses = Object.entries(requestCodegen(swaggerSource.paths, isV3, options))
+
+  const { models, enums } = isV3
+    ? componentsCodegen(swaggerSource.components)
+    : definitionsCodeGen(swaggerSource.definitions)
+
   // TODO: next next next time
   // if (options.multipleFileMode) {
   if (false) {
-
-    Object.entries(requestCodegen(swaggerSource.paths)).forEach(([className, requests]) => {
+    Object.entries(requestCodegen(swaggerSource.paths, isV3, options)).forEach(([className, requests]) => {
       let text = ''
       requests.forEach(req => {
-
-        const reqName = options.methodNameMode == "operationId"
-          ? req.operationId
-          : req.name
+        const reqName = options.methodNameMode == 'operationId' ? req.operationId : req.name
         text += requestTemplate(reqName, req.requestSchema, options)
       })
       text = serviceTemplate(className + options.serviceNameSuffix, text)
@@ -73,9 +85,7 @@ export async function codegen(params: ISwaggerOptions) {
     const { models, enums } = definitionsCodeGen(swaggerSource.definitions)
     let defsString = ''
     Object.values(enums).forEach(item => {
-      const text = item.value
-        ? enumTemplate(item.value.name, item.value.enumProps, 'Enum')
-        : item.content || ''
+      const text = item.value ? enumTemplate(item.value.name, item.value.enumProps, 'Enum') : item.content || ''
 
       // const fileDir = path.join(options.outputDir || '', 'definitions')
       // writeFile(fileDir, item.name + '.ts', format(text, options))
@@ -83,25 +93,33 @@ export async function codegen(params: ISwaggerOptions) {
     })
 
     Object.values(models).forEach(item => {
-      const text = params.modelMode === 'interface'
-        ? interfaceTemplate(item.value.name, item.value.props, [], params.strictNullChecks)
-        : classTemplate(item.value.name, item.value.props, [], params.strictNullChecks, options.useClassTransformer)
+      const text =
+        params.modelMode === 'interface'
+          ? interfaceTemplate(item.value.name, item.value.props, [], params.strictNullChecks)
+          : classTemplate(
+            item.value.name,
+            item.value.props,
+            [],
+            params.strictNullChecks,
+            options.useClassTransformer,
+            options.generateValidationModel
+          )
       // const fileDir = path.join(options.outputDir || '', 'definitions')
       // writeFile(fileDir, item.name + '.ts', format(text, options))
       defsString += text
     })
     writeFile(options.outputDir || '', 'index.defs.ts', format(defsString, options))
-  }
-  else if (options.include && options.include.length > 0) {
+  } else if (options.include && options.include.length > 0) {
+    // 接口过滤入口
     let reqSource = ''
     let defSource = ''
-    let requestClasses = Object.entries(requestCodegen(swaggerSource.paths))
-    const { models, enums } = definitionsCodeGen(swaggerSource.definitions)
 
     let allModel = Object.values(models)
     // console.log(allModel)
     let allEnum = Object.values(enums)
     let allImport: string[] = []
+
+    // 处理接口
 
     options.include.forEach(item => {
       let includeClassName = ''
@@ -113,15 +131,12 @@ export async function codegen(params: ISwaggerOptions) {
           includeClassName = k
           includeRequests = (<IInclude>item)[k]
         }
-
       }
       for (let [className, requests] of requestClasses) {
         if (pascalcase(includeClassName) !== className) continue
         let text = ''
         for (let req of requests) {
-          const reqName = options.methodNameMode == "operationId"
-            ? req.operationId
-            : req.name
+          const reqName = options.methodNameMode == 'operationId' ? req.operationId : req.name
           if (includeRequests) {
             if (includeRequests.includes(reqName)) {
               text += requestTemplate(reqName, req.requestSchema, options)
@@ -141,26 +156,36 @@ export async function codegen(params: ISwaggerOptions) {
       }
     })
 
+    // 处理类和枚举
+
     allModel.forEach(item => {
       if (allImport.includes(item.name)) {
-        const text = params.modelMode === 'interface'
-          ? interfaceTemplate(item.value.name, item.value.props, [], params.strictNullChecks)
-          : classTemplate(item.value.name, item.value.props, [], params.strictNullChecks, options.useClassTransformer)
+        const text =
+          params.modelMode === 'interface'
+            ? interfaceTemplate(item.value.name, item.value.props, [], params.strictNullChecks)
+            : classTemplate(
+              item.value.name,
+              item.value.props,
+              [],
+              params.strictNullChecks,
+              options.useClassTransformer,
+              options.generateValidationModel
+            )
         defSource += text
       }
     })
 
     allEnum.forEach(item => {
       if (allImport.includes(item.name)) {
-        let text = ''; 
-        if(item.value){
-          if(item.value.type == 'string'){
-            text = enumTemplate(item.value.name, item.value.enumProps, options.enumNamePrefix);
-          }else{
+        let text = ''
+        if (item.value) {
+          if (item.value.type == 'string') {
+            text = enumTemplate(item.value.name, item.value.enumProps, options.enumNamePrefix)
+          } else {
             text = typeTemplate(item.value.name, item.value.enumProps, options.enumNamePrefix)
           }
-        }else{
-          text = item.content || '';
+        } else {
+          text = item.content || ''
         }
 
         defSource += text
@@ -169,46 +194,50 @@ export async function codegen(params: ISwaggerOptions) {
 
     apiSource += reqSource + defSource
     writeFile(options.outputDir || '', options.fileName || '', format(apiSource, options))
-  }
-  else {
+  } else {
+    // 常规入口
     try {
-
-      Object.entries(requestCodegen(swaggerSource.paths)).forEach(([className, requests]) => {
+      // 处理接口
+      requestClasses.forEach(([className, requests]) => {
         let text = ''
         requests.forEach(req => {
-
-          const reqName = options.methodNameMode == "operationId"
-            ? req.operationId
-            : req.name
+          const reqName = options.methodNameMode == 'operationId' ? req.operationId : req.name
           text += requestTemplate(reqName, req.requestSchema, options)
         })
         text = serviceTemplate(className + options.serviceNameSuffix, text)
         apiSource += text
       })
 
-      const { models, enums } = definitionsCodeGen(swaggerSource.definitions)
+      // 处理类和枚举
 
       Object.values(models).forEach(item => {
-        const text = params.modelMode === 'interface'
-          ? interfaceTemplate(item.value.name, item.value.props, [], params.strictNullChecks)
-          : classTemplate(item.value.name, item.value.props, [], params.strictNullChecks, options.useClassTransformer)
+        const text =
+          params.modelMode === 'interface'
+            ? interfaceTemplate(item.value.name, item.value.props, [], params.strictNullChecks)
+            : classTemplate(
+              item.value.name,
+              item.value.props,
+              [],
+              params.strictNullChecks,
+              options.useClassTransformer,
+              options.generateValidationModel
+            )
         apiSource += text
       })
 
       Object.values(enums).forEach(item => {
-        let text = ''; 
-        if(item.value){
-          if(item.value.type == 'string'){
-            text = enumTemplate(item.value.name, item.value.enumProps, options.enumNamePrefix);
-          }else{
+        let text = ''
+        if (item.value) {
+          if (item.value.type == 'string') {
+            text = enumTemplate(item.value.name, item.value.enumProps, options.enumNamePrefix)
+          } else {
             text = typeTemplate(item.value.name, item.value.enumProps, options.enumNamePrefix)
           }
-        }else{
-          text = item.content || '';
+        } else {
+          text = item.content || ''
         }
         apiSource += text
       })
-
 
       writeFile(options.outputDir || '', options.fileName || '', format(apiSource, options))
     } catch (error) {
@@ -217,21 +246,20 @@ export async function codegen(params: ISwaggerOptions) {
     }
   }
   if (fs.existsSync('./cache_swagger.json')) {
-    fs.unlinkSync('./cache_swagger.json');
+    fs.unlinkSync('./cache_swagger.json')
   }
   console.timeEnd('finish')
   if (err) {
-    throw err;
+    throw err
   }
 }
 
-
 function writeFile(fileDir: string, name: string, data: any) {
   if (!fs.existsSync(fileDir)) {
-    fs.mkdirSync(fileDir);
+    fs.mkdirSync(fileDir)
   }
   const filename = path.join(fileDir, name)
-  console.log('filename', filename);
+  console.log('filename', filename)
   fs.writeFileSync(filename, data)
 }
 
@@ -242,12 +270,13 @@ function format(text: string, options: ISwaggerOptions) {
   }
   console.log('use default formatter')
   return prettier.format(text, {
-    "printWidth": 120,
-    "tabWidth": 2,
-    "parser": "typescript",
-    "trailingComma": "none",
-    "jsxBracketSameLine": false,
-    "semi": true,
-    "singleQuote": true
+    printWidth: 120,
+    tabWidth: 2,
+    parser: 'typescript',
+    trailingComma: 'none',
+    jsxBracketSameLine: false,
+    semi: true,
+    singleQuote: true
   })
+  // return text
 }