+++ src/requestCodegen/getResponseType.ts
@@ -1,9 +1,10 @@
-import { refClassName, toBaseType } from '../utils'
 import { IRequestMethod } from '../swaggerInterfaces'
+import { refClassName, toBaseType } from '../utils'
+
 /**
  * 获取请求的返回类型
  */
-export function getResponseType(reqProps: IRequestMethod): { responseType: string, isRef: boolean } {
+export function getResponseType(reqProps: IRequestMethod, isV3: boolean): { responseType: string, isRef: boolean } {
   // It does not allow the schema defined directly, but only the primitive type is allowed.
   let result: string = 'any'
   let isRef = false
@@ -12,17 +13,26 @@ export function getResponseType(reqProps: IRequestMethod): { responseType: strin
   if (!successStatusCode) {
     return { responseType: result, isRef }
   }
-  const resSchema =
-    reqProps.responses[successStatusCode] && reqProps.responses[successStatusCode].schema
-      ? reqProps.responses[successStatusCode].schema
-      : null
+  let resSchema = null
+  if (reqProps.responses[successStatusCode]) {
+    if (isV3 === true) {
+      if (
+        reqProps.responses[successStatusCode].content &&
+        reqProps.responses[successStatusCode].content['application/json'] &&
+        reqProps.responses[successStatusCode].content['application/json'].schema
+      )
+        resSchema = reqProps.responses[successStatusCode].content['application/json'].schema
+    } else {
+      if (reqProps.responses[successStatusCode].schema) resSchema = reqProps.responses[successStatusCode].schema
+    }
+  }
 
   if (!resSchema) {
     return { responseType: result, isRef }
   }
 
   let checkType = resSchema.type
-  let format = resSchema.format;
+  let format = resSchema.format
   // 如果是数组
   if (checkType === 'array' || resSchema.items) {
     if (resSchema.items.$ref) {