+++ src/swaggerInterfaces.ts
@@ -1,9 +1,11 @@
 export interface ISwaggerSource {
-  swagger: string
+  swagger?: string | undefined,
+  openapi?: string | undefined,
   info: string
   paths: IPaths
   securityDefinitions: string
   definitions: IDefinitions
+  components: IComponents
   externalDocs: string
 }
 
@@ -23,15 +25,36 @@ export interface IRequestMethod {
   consumes: string[]
   produces: string[]
   parameters: IParameter[]
+  requestBody: IRequestBody,
   responses: {
     [key: string]: {
       description: string
+      // v2
       schema: {
         '$ref': string,
         'type'?: string,
         'items'?: IParameterItems,
         'format'?: string,
+      },
+      // v3
+      content: {
+        [key: string]: {
+          schema: {
+            '$ref': string,
+            'type'?: string,
+            'items'?: IParameterItems,
+            'format'?: string,
+          }
+        }
       }
+    }
+  }
+}
+
+export interface IRequestBody {
+  content: {
+    [key: string]: {
+      schema: ISchema
     }
   }
 }
@@ -87,6 +110,38 @@ export interface IDefinitionProperty {
   format: string
   maxLength: number
   $ref: string
+  allOf: IDefinitionProperties
   items: IDefinitionProperty
   description: string
+}
+
+export interface IComponents {
+  schemas: {
+    [key: string]: IDefinition
+  }
+}
+
+export interface ISchema {
+  '$ref': string
+  'type'?: string
+  'items'?: IParameterItems
+  'format'?: string
+}
+
+// 字典类型，对应java map或者.NET Dictionary
+
+export interface IDictionary<TKey, TValue = any> {
+  [key: string]: TValue;
+}
+
+export interface IDictionary<TKey, TValue = any> {
+  [key: number]: TValue;
+}
+
+export class Dictionary<TKey, TValue> implements IDictionary<TKey, TValue> {
+  [key: string]: TValue;
+}
+
+export class Map<TKey, TValue> implements IDictionary<TKey, TValue> {
+  [key: string]: TValue;
 }